god oject cleanup 2
